Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/svg/extensibility/foreignObject/containing-block-expected.txt


FAIL position assert_equals: offsetLeft expected 5 but got 15
